requirements.txt
Django==3.2.13
django-debug-toolbar==3.5.0
flake8-django==1.1.2
flake8-isort==4.1.1
Pillow==9.2.0
django-crispy-forms==1.14.0
django-modeltranslation==0.18.4
djangorestframework~=3.13.1
celery~=5.2.7
redis~=4.3.4